requirements.txt
asyncpg==0.28.0
web3==6.11.0
python-dotenv==1.0.0
backoff==2.2.1
pydantic==2.4.2
httpx==0.25.0